[UPDATE/1123] Door 오브젝트에 녹화 기능 업데이트

+++ KDH_DX2D_KZ/GameEngineContents/LaserGroup.cpp
@@ -23,7 +23,7 @@ void LaserGroup::InitLaserGroupData(int _LaserCount, float4 _InitPos, float _XDi
 
 	if (true == IsUseMoving)
 	{
-		GroupLastCount = AllLaser.size();
+		GroupLastCount = static_cast<int>(AllLaser.size());
 	}
 }
